.paa/requirements/requirements_package_auto_assembler.txt
attrs>=22.2.0
pip_audit==2.7.3
pandas
mkdocs-material>=9.5.30
attrsx
twine>=5.1.1
streamlit>=1.39.0
uvicorn
nbformat
packaging
jupyter>=1.1.1
mkdocs>=1.6.0,<2.0
fastapi
click==8.1.7
mkdocs-mermaid2-plugin>=1.2.1
numpy==1.26.0
mcp
stdlib-list
wheel>=0.44.0
setuptools>=78.1.1
nbconvert>=7.16.4
pyyaml
requests
